hardware/ril: Fix rild crash when phone app gets killed continuously
Whenever phone app gets killed, the listen events are getting piled up in
watch table, thus not allowing further fd command events to be registered
which is resulting in rild crash. As part of DSDS changes, listen event was
made persistent. So, it needs to be deleted and then added to the watch list
whenever connection is reset.

Change-Id: I8fcd5aa89372fb1ae19edd93fe12fa340f610656
CRs-Fixed: 263050